requirements.txt
python-socketio == 5.9.0
Quart == 0.19.3
hypercorn == 0.14.4
aiofiles == 23.2.1
azure-cognitiveservices-speech == 1.32.1
